requirements.txt
# requirements.txt
streamlit==1.28.1
pandas==2.1.3
sqlalchemy==2.0.23
plotly==5.17.0
openai==1.3.5
sentence-transformers==2.2.2
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.17.1
beautifulsoup4==4.12.2
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.0
bleach==6.1.0
psutil==5.9.6
redis==5.0.1
celery==5.3.4
uvicorn==0.24.0
fastapi==0.104.1
loguru==0.7.2
schedule==1.2.0
croniter==2.0.1
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
jinja2==3.1.2
markdown==3.5.1
python-dateutil==2.8.2
pytz==2023.3
validators==0.22.0
pillow==10.1.0
numpy==1.26.4
matplotlib==3.8.2
seaborn==0.13.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0
